requirements.txt
streamlit>=1.28.0
crewai>=0.1.0
openai>=1.0.0
playwright>=1.39.0
pydantic>=2.0.0
langchain-community>=0.0.20
assemblyai>=0.20.0
notion-client>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
langchain-google-community[gmail]
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
moviepy>=1.0.3
pydantic-ai>=0.0.12
langfuse>=2.0.0
opencv-python>=4.10.0
pyautogui>=0.9.54
sounddevice
psutil
plotly>=6.0.0
pyaudio>=0.2.11
pydub>=0.25.1



